Feat: implement Speedometer (GPS speed tracking)
- TDD: 16 unit tests for SpeedometerViewModel (RED→GREEN)
- SpeedometerViewModel: m/s→km/h conversion, max speed tracking, unit switching (km/h, mph, knots)
- SpeedometerScreen: FusedLocationProvider GPS integration, gauge UI, unit chips, reset
- ToolRegistry: registered speedometer with ACCESS_FINE_LOCATION permission
- Dependencies: play-services-location 21.3.0, kotlinx-coroutines-play-services 1.8.1
- Fixes: LuxMeterScreen type issues, MetalDetectorScreen Composable hoisting, SoundMeterViewModel abs/Byte

+++ app/src/main/java/com/armyknife/tools/features/sensor/SoundMeterViewModel.kt
@@ -20,7 +20,7 @@ data class SoundMeterUiState(
     val peakDb: Float = 0f,
     val isRecording: Boolean = false,
     val levelCategory: String = "Silent",
-    val rawAmplitude: Int = 0,
+    val rawAmplitude: Double = 0.0,
     val history: List<Float> = emptyList()
 )
 
@@ -91,11 +91,11 @@ class SoundMeterViewModel : ViewModel() {
                 if (read > 0) {
                     // Compute RMS
                     var sum = 0.0
-                    var maxAbs = 0
+                    var maxAbs = 0.0
                     for (i in 0 until read) {
                         val sample = buffer[i].toDouble()
                         sum += sample * sample
-                        val abs = kotlin.math.abs(buffer[i])
+                        val abs = kotlin.math.abs(buffer[i].toDouble())
                         if (abs > maxAbs) maxAbs = abs
                     }
                     val rms = kotlin.math.sqrt(sum / read)
